refactor: add enum moved to module
